Extract analysis move tree out of guimoveslist into an opt-in renderer
